[FEAT] : ⚙️ 메인페이지와 연동
